% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{bar_styles}
\alias{bar_styles}
\title{bar_styles
Initialize a set of parameter values for use with the
bar_timelines function}
\usage{
bar_styles()
}
\value{
A data frame with default parameter values
}
\description{
bar_styles
Initialize a set of parameter values for use with the
bar_timelines function
}
\examples{
column <- bar_styles()
}
