% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-path.R
\name{build_network}
\alias{build_network}
\title{Build the citation network}
\usage{
build_network(M)
}
\arguments{
\item{M}{A bibliometrix M data frame.}
}
\value{
An igraph citation network.
}
\description{
It will call other functions to build the citation
net and simplify it into an acyclic graph.
}
