% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-path.R
\name{family_attr_comb}
\alias{family_attr_comb}
\title{Combines attribute for merged paper nodes}
\usage{
family_attr_comb(x)
}
\arguments{
\item{x}{A vector with the titles of the papers.}
}
\value{
The combined titles.
}
\description{
When simplifying networks by merging nodes/papers, this
is the function used to combine the paper titles in one.
}
