% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-path.R
\name{find_cycles}
\alias{find_cycles}
\title{Find cycles of a specified length in a graph}
\usage{
find_cycles(g, size)
}
\arguments{
\item{g}{The igraph network.}

\item{size}{The length of the loops to find.}
}
\value{
A list of vectors, each vector providing the ids of the nodes in the loop.
}
\description{
Helper function for making networks acyclical. It
finds loops of a given order in an igraph network.
}
