% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblio.R
\name{get_CR_dois}
\alias{get_CR_dois}
\title{Builds parts of queries}
\usage{
get_CR_dois(M)
}
\arguments{
\item{M}{A bibliometrix M data frame.}
}
\value{
A character vector with the queries for the cited references.
}
\description{
Helper function for making Web of Science queries using
the DOIs of the cited references.
}
