% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpys.R
\name{identify_rpys_milestones}
\alias{identify_rpys_milestones}
\title{Identifies milestones through RPYS papers}
\usage{
identify_rpys_milestones(M)
}
\arguments{
\item{M}{A bibliometric data frame (from bibliometrix).}
}
\value{
A list containing a data frame of the documents and the RPYS plot.
}
\description{
Given a bibliometric data frame, it identifies the cited reference years
and builds the RPYS graph. It then finds the peak years and extracts the
most cited references in each of those peaks. These are the documents
returned with the graph (the graph has the peaks highlighted).
}
