% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpys.R
\name{is_peak}
\alias{is_peak}
\title{Finds peaks in time series data}
\usage{
is_peak(variable)
}
\arguments{
\item{variable}{A numeric vector where peaks will be identified.}
}
\value{
A logical vector matching the variable argument, identifying the peaks
}
\description{
Finds local maxima in a time series vector. It will return a matching
logical vector with TRUE where the direction of growth changes, i.e.
on the peaks.
}
