% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{lolli_styles}
\alias{lolli_styles}
\title{lolli_styles
Initialize a set of parameter values for use with the
milestones function}
\usage{
lolli_styles()
}
\value{
A dataframe with default parameter values
}
\description{
lolli_styles
Initialize a set of parameter values for use with the
milestones function
}
\examples{
column <- lolli_styles()
}
