% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-path.R
\name{make_hist_citation_net}
\alias{make_hist_citation_net}
\title{Matches citations to build citation net}
\usage{
make_hist_citation_net(M, min.citations = 1)
}
\arguments{
\item{M}{The bibliometrix data frame.}

\item{min.citations}{The minimum number of received citations for the paper to be included.}
}
\value{
A list with the citation information and metadata.
}
\description{
Matches the names in the cited references field to the
short titles to identify papers in the dataset that cited
each other. It's a version of the original bibliometrix
version with a few steps omitted for performance.
}
