% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-path.R
\name{make_loops_into_families}
\alias{make_loops_into_families}
\title{Merges papers which are in a citation loop}
\usage{
make_loops_into_families(loops, g)
}
\arguments{
\item{loops}{A list of loops, each a vector with node ids.}

\item{g}{The citation igraph network.}
}
\value{
A data frame with the information extracted.
}
\description{
Merges the papers passed to it into one node. This is used
when removing loops to make the citation net acyclic. If A
cites B and B cites A, this will become node AB, with their
outgoing and ingoing edges pooled together.
}
