% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblio.R
\name{make_wos_query}
\alias{make_wos_query}
\title{Builds queries from cited references}
\usage{
make_wos_query(M, next_name)
}
\arguments{
\item{M}{A bibliometrix M data frame.}

\item{next_name}{The name you want to give to the generation (e.g. G1).}
}
\value{
A list with the cited reference years, counted as indicated.
}
\description{
For a set of Web of Science records, this function builds and saves
text files containing queries that can be pasted on Web of Science to
obtain the records for the references cited by the first set of records
(i.e. the next generation of cited references). The query is built twice,
using DOIs and then author names and journals. These are redundant to
obtain a better recall in the search.
}
