% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{milestones}
\alias{milestones}
\title{milestones
Builds charts using "ggplot2" with lollipop shape symbols
marking events on a timeline}
\usage{
milestones(
  datatable = data,
  styles = column,
  add_row_above = 0.5,
  add_row_below = 0.5
)
}
\arguments{
\item{datatable}{A data frame with columns for events and dates}

\item{styles}{A data frame with the style parameters}

\item{add_row_above}{A binary variable to add spacing at the top}

\item{add_row_below}{A binary variable to add spacing below}
}
\value{
A "ggplot2" object ready for plotting
}
\description{
milestones
Builds charts using "ggplot2" with lollipop shape symbols
marking events on a timeline
}
\examples{
milestones(data=data,styles=column,add_row_below=TRUE)
}
