% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblio.R
\name{read_biblio_data}
\alias{read_biblio_data}
\title{Load a bibliometric data frame}
\usage{
read_biblio_data(filenames = NULL, path = NULL)
}
\arguments{
\item{filenames}{A vector with the filenames.}

\item{path}{Full path to the folder containing the bibliometric data.}
}
\value{
An "M" data frame, as loaded and converted through bibliometrix.
}
\description{
Obtains the "M" data frame using bibliometrix to read a vector of files
or all text files in the given folder if the argument path is given.
Files must be those downloaded from Web of Science, in the plain text format.
}
