% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-path.R
\name{read_network}
\alias{read_network}
\title{Pajek file reader: obtains the main path saved from Pajek}
\usage{
read_network(filename, net)
}
\arguments{
\item{filename}{The Pajek exported file with the main path network.}

\item{net}{The igraph full network.}
}
\value{
An igraph network with only the main path nodes.
}
\description{
The main path file will be exported from Pajek. This function
reads it back into R as an igraph network built by subsetting
the full network.
}
\details{
Code is adapted from function *read_net*, from Jonathan H. Morgan (2019)
http://mrvar.fdv.uni-lj.si/pajek/R/RMorgan.htm
}
