% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-path.R
\name{save_network}
\alias{save_network}
\title{Saves the citation network}
\usage{
save_network(NET, filename)
}
\arguments{
\item{NET}{The igraph object to save.}

\item{filename}{Where to save the files. Extensions will be appended.}
}
\value{
An igraph citation network.
}
\description{
Saves a citation network in GML and also in a
.net format readable by the Pajek software.
}
