% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-path.R
\name{save_papers}
\alias{save_papers}
\title{Save list of main path papers}
\usage{
save_papers(RPYS_MS, filename)
}
\arguments{
\item{NET}{The main path network, as an igraph object.}

\item{output_path}{Path to a directory to save the main path files.}

\item{M}{A bibliometrix M data frame.
#' @return An bibliometrix data frame with only the main path nodes.}
}
\description{
Saves the igraph object of a network as a GML file and also
saves the list of papers in the main path, as well as the
full Web of Science record for the papers in the main path
(if given a bibliometric data frame).
}
