% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpys.R
\name{save_papers_rpys}
\alias{save_papers_rpys}
\title{Save list of RPYS milestones papers}
\usage{
save_papers_rpys(RPYS_MS, filename)
}
\arguments{
\item{RPYS_MS}{The object returned by the *identify_rpys_milestones* function.}

\item{filename}{Filename to save the list of milestone references.}
}
\description{
Saves the list of papers identified through the RPYS algorithm
}
